Add files via upload
A copy right based technique used for adding extra security to a images and copy which is copy righted, by adding a specific watermark by a owner provides a file an Extra protection, where those files cannot be reproduced and a part of content, when reproduces specifics the watermarking or water marked content is not visible. By adding this protection the Originality between Copy and Original can be easy  determined. For more contact : Sonusantho@gmail.com